refactor(earntri,select): replace sort buttons by select. Remove ordering by ASC
