+++ src/components/EarnTriSortAndFilter/EarnTriSortAndFilterContainer.tsx
@@ -4,7 +4,7 @@ import { Text } from 'rebass'
 import Toggle from '../Toggle'
 import { useToggleFilterActiveFarms } from '../../state/user/hooks'
 
-import Select from '../Select'
+import Select, { OptionProps } from '../Select'
 
 import {
   StyledSearchInput,
@@ -14,21 +14,26 @@ import {
   StyledSortOption,
   StyledArrowContainer
 } from '../../pages/EarnTri/EarnTri.styles'
-import { SortingType } from './EarnTriSortAndFilterSortingType'
+
 import SortingArrow from './SortingArrow'
+import { SortingType } from '../../pages/EarnTri/useFarmsSortAndFilter'
+
+const SORT_OPTIONS = [
+  { label: 'PoolType', value: 'default' },
+  { label: 'Liquidity', value: 'liquidity' },
+  { label: 'APR', value: 'totalApr' }
+]
 
 export default function EarnTriSortAndFilterContainer({
   activeFarmsFilter,
   handleSort,
   isSortDescending,
-  onInputChange,
-  sortBy
+  onInputChange
 }: {
   activeFarmsFilter: boolean
   handleSort: (sortingType: SortingType) => void
   isSortDescending: boolean
   onInputChange: (event: React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>) => void
-  sortBy: SortingType
 }) {
   const { t } = useTranslation()
   const toggleActiveFarms = useToggleFilterActiveFarms()
@@ -44,33 +49,11 @@ export default function EarnTriSortAndFilterContainer({
 
           <Toggle id="toggle-user-farms-toggle" isActive={activeFarmsFilter} toggle={toggleActiveFarms} />
         </StyledToggleContainer>
-        <Select
-          options={[
-            { label: 'All pools', value: 'allPools' },
-            { label: 'Stable Pools', value: 'stable' },
-            { label: 'Dual Reward pools', value: 'dualRewards' },
-            { label: 'Tri only pools', value: 'triOnly' },
-            { label: 'Ecosystem pools', value: 'ecosystem' },
-            { label: 'Legacy pools', value: 'legacy' }
-          ]}
-        />
         <StyledSortContainer>
-          <Text fontWeight={400} fontSize={16}>
-            Sort by:{' '}
-            <StyledSortOption onClick={() => handleSort(SortingType.liquidity)}>
-              {SortingType.liquidity}
-              <StyledArrowContainer>
-                {sortBy === SortingType.liquidity && <SortingArrow isDescending={isSortDescending} />}
-              </StyledArrowContainer>
-            </StyledSortOption>
-            |
-            <StyledSortOption style={{ marginLeft: '1rem' }} onClick={() => handleSort(SortingType.totalApr)}>
-              {SortingType.totalApr}
-              <StyledArrowContainer>
-                {sortBy === SortingType.totalApr && <SortingArrow isDescending={isSortDescending} />}
-              </StyledArrowContainer>
-            </StyledSortOption>
+          <Text fontWeight={400} fontSize={16} marginRight={20}>
+            Sort by:
           </Text>
+          <Select options={SORT_OPTIONS} onOptionChange={(option: OptionProps) => handleSort(option.value)} />
         </StyledSortContainer>
       </StyledFiltersContainer>
     </>
