--- src/components/EarnTriSortAndFilter/EarnTriSortAndFilterSortingType.tsx
@@ -1,5 +1,0 @@
-export enum SortingType {
-  liquidity = 'Liquidity',
-  totalApr = 'Total APR',
-  default = 'Default'
-}
